requirements.txt
flask
sqlalchemy
flask-sqlalchemy
flask-wtforms
Flask-Migrate
# MySQL
# PyMySQL
# Postgres
# psycopg2
# MSSQL
# pyodbc
# Oracle
# cx_Oracle